- Fixed browser title if theme has title-tag support - Fixed the video player issue for activity load more - Added filter for media upload button ID - Check whether upload media button exist or not in rtMedia.backbone.js for activity page - Fix rtTab js issue - Fixed BuddyPress group privacy issue - Fixed 404 error for titles on bbPress  templates - CSS fixes - Removed extra params while fetching media gallery after uploading media - Fix comments issue

+++ app/helper/RTMediaCommentNotification.php
@@ -111,8 +111,11 @@ class RTMediaCommentNotification extends RTMediaNotification {
 	 */
 	function remove_comment_notification( $comment_id ) {
 		$comment_notification_id = (int) get_comment_meta( $comment_id, 'comment_notification_id', true );
-		BP_Notifications_Notification::delete( array( 'id' => $comment_notification_id ) );
-		delete_comment_meta( $comment_id, 'comment_notification_id' );
+
+		if ( $comment_notification_id > 0 ) {
+			BP_Notifications_Notification::delete( array( 'id' => $comment_notification_id ) );
+			delete_comment_meta( $comment_id, 'comment_notification_id' );
+		}
 	}
 }
